+++ src/server/package.json
@@ -12,14 +12,14 @@
     "clean": "rm -rf dist"
   },
   "dependencies": {
-    "@prisma/client": "^4.4.0",
+    "@prisma/client": "^5.2.0",
     "@trpc/server": "^10.7.0",
     "cors": "^2.8.5",
     "dotenv": "^16.0.3",
     "dotenv-cli": "^6.0.0",
     "express": "^4.18.1",
     "nanoid": "^3.3.4",
-    "prisma": "^4.4.0",
+    "prisma": "^5.2.0",
     "ws": "^8.9.0",
     "zod": "^3.19.1"
   },
