--- src/server/src/live-room/results.ts
@@ -1,13 +1,0 @@
-export interface CandidateWithVotes {
-  id: string;
-  name: string;
-  votes: number;
-}
-
-export interface SingleVoteResultsView {
-  type: 'SingleVote';
-  results: CandidateWithVotes[];
-}
-
-// TODO: Add more types
-export type ResultsView = { abstained: number } & SingleVoteResultsView;
