+++ src/server/src/room/interaction/admin.ts
@@ -1,5 +1,5 @@
 import { roomVoterNotifications, roomWaitingListNotifications, userWaitingListNotifications } from '../../live';
-import { VoterState } from '../../live-room/live-states';
+import { VoterState } from '../../live/states';
 import type { CreateQuestionParams } from '../types';
 import type { RoomUsersList } from './db/users';
 import { RoomUserState } from './db/users';
