+++ src/server/src/routers/room-vote.ts
@@ -1,8 +1,8 @@
 import { observable } from '@trpc/server/observable';
 import { z } from 'zod';
 
-import type { VoterState } from '../live-room/live-states';
-import { questionResponse } from '../live-room/question';
+import { questionResponse } from '../live/question';
+import type { VoterState } from '../live/states';
 import { operations } from '../room';
 import { publicProcedure, router } from '../trpc';
