+++ src/web/src/pages/room/RoomResultsListPage.tsx
@@ -1,9 +1,9 @@
-import type { PublicStaticRoomData } from '@server/rooms';
+import type { RoomPublicInfo } from '@server/room/types';
 import { ResultsViewer } from 'components/ResultsViewer';
 import { Heading, PageContainer, Question } from 'components/styles';
 import { trpc } from 'utils/trpc';
 
-export function RoomResultsListPage(props: { roomId: string; room: PublicStaticRoomData }) {
+export function RoomResultsListPage(props: { roomId: string; room: RoomPublicInfo }) {
   const roomResults = trpc.room.getResults.useQuery({ roomId: props.roomId });
 
   return (
@@ -15,8 +15,8 @@ export function RoomResultsListPage(props: { roomId: string; room: PublicStaticR
         ) : (
           roomResults.data.map((result) => (
             <div key={result.questionId}>
-              <Question className="mb-4">{result.name}</Question>
-              <ResultsViewer results={result.result} />
+              <Question className="mb-4">{result.question}</Question>
+              <ResultsViewer results={result.results} />
             </div>
           ))
         )}
